fix immutability test.
This commit fixes an issue where an object that does
implement __hash__ but has it raise an exception
internally could potentially be considered immutable.

Currently the only case I'm aware of is tuple that will
raise an exception when __hash__ is called if one of
its values does not implement hash.

But fortunately this didn't really cause an issue because
tuples are always converted into FLists regardless of their
hashability.